fix always fullscreen mode
The window layout was always being set to FULLSCREEN mode to workaround a flickering status bar bug, but it did not respect the app preference and return to FORCE_NOT_FULLSCREEN when applicable.  This had a side effect of breaking the adjustResize windowSoftInputMode setting, where the soft keyboard being shown would not resize the viewport.